+++ Package.swift
@@ -6,7 +6,7 @@ import PackageDescription
 let package = Package(
     name: "swift-nio-email",
     products: [
-        .library(name: "NIOIMAP", targets: ["NIOIMAP", "NIOIMAPCore"])
+        .library(name: "NIOIMAP", targets: ["NIOIMAPCore"])
     ], dependencies: [
         .package(url: "https://github.com/apple/swift-nio", from: "2.16.0"),
         .package(url: "https://github.com/apple/swift-nio-extras", from: "1.4.0"),
@@ -14,37 +14,26 @@ let package = Package(
         .package(url: "https://github.com/apple/swift-log", from: "1.2.0")
     ],
     targets: [
-        .target(
-            name: "NIOIMAP",
-            dependencies: ["NIOIMAPCore"]
-        ),
-        .target(
-            name: "NIOIMAPCore",
-            dependencies: ["NIO"]
-        ),
-        .target(
-            name: "CLI",
-            dependencies: ["CLILib"]
-        ),
-        .target(
-            name: "CLILib",
-            dependencies: ["NIO", "NIOSSL", "NIOIMAP", "Logging"]
-        ),
-        .target(
-            name: "Proxy",
-            dependencies: ["NIOIMAP", "NIO", "NIOSSL"]
-        ),
+        
+        .target(name: "NIOIMAP", dependencies: ["NIOIMAPCore"]),
         .testTarget(
             name: "NIOIMAPTests",
             dependencies: ["NIOIMAP", "NIOIMAPCore", "NIO", "NIOTestUtils"]
         ),
+        
+        .target(name: "NIOIMAPCore", dependencies: ["NIO"]),
         .testTarget(
             name: "NIOIMAPCoreTests",
             dependencies: ["NIOIMAPCore", "NIO", "NIOTestUtils"]
         ),
+        
+        .target(name: "CLI", dependencies: ["CLILib"]),
+        .target(name: "CLILib",dependencies: ["NIO", "NIOSSL", "NIOIMAP", "Logging"]),
         .testTarget(
             name: "CLILibTests",
             dependencies: ["CLILib", "NIO", "NIOTestUtils"]
         ),
+        
+        .target(name: "Proxy", dependencies: ["NIOIMAP", "NIO", "NIOSSL"]),
     ]
 )
